backend/requirements.txt
# Core Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
html2text==2020.1.16
python-dotenv==1.0.0
pandas==2.1.4
pytz==2023.3

# Optional Dependencies
colorama==0.4.6
validators==0.22.0
celery==5.3.4
redis==5.0.1

# Development Dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
